init_grafana_plugin fixture を Grafana テスト専用として同モジュール内に移動する
